requirements.txt
gtts
GoogleBard
playsound
portaudio
SpeechRecognition
pygobject
pipwin
